Added explicit casts and changed socket holding variables to be 64bits compatible (compiled with Visual Studio). Before it will have tons of warnings because of implicit casts. Some datatypes had to be changed, some explicit casts added.

+++ src/bson.c
@@ -112,7 +112,7 @@ MONGO_EXPORT int bson_size( const bson *b ) {
     return i;
 }
 
-MONGO_EXPORT int bson_buffer_size( const bson *b ) {
+MONGO_EXPORT size_t bson_buffer_size( const bson *b ) {
     return (b->cur - b->data + 1);
 }
 
@@ -312,7 +312,7 @@ MONGO_EXPORT bson_bool_t bson_iterator_more( const bson_iterator *i ) {
 }
 
 MONGO_EXPORT bson_type bson_iterator_next( bson_iterator *i ) {
-    int ds;
+    size_t ds;
 
     if ( i->first ) {
         i->first = 0;
@@ -625,7 +625,7 @@ void bson_append64( bson *b, const void *data ) {
 }
 
 int bson_ensure_space( bson *b, const int bytesNeeded ) {
-    int pos = b->cur - b->data;
+    size_t pos = b->cur - b->data;
     char *orig = b->data;
     int new_size;
 
@@ -654,7 +654,7 @@ int bson_ensure_space( bson *b, const int bytesNeeded ) {
 }
 
 MONGO_EXPORT int bson_finish( bson *b ) {
-    int i;
+    size_t i;
 
     if( b->err & BSON_NOT_UTF8 )
         return BSON_ERROR;
@@ -683,7 +683,7 @@ MONGO_EXPORT void bson_destroy( bson *b ) {
 }
 
 static int bson_append_estart( bson *b, int type, const char *name, const int dataSize ) {
-    const int len = strlen( name ) + 1;
+    const int len = (int)strlen( name ) + 1;
 
     if ( b->finished ) {
         b->err |= BSON_ALREADY_FINISHED;
@@ -764,15 +764,15 @@ int bson_append_string_base( bson *b, const char *name,
 }
 
 MONGO_EXPORT int bson_append_string( bson *b, const char *name, const char *value ) {
-    return bson_append_string_base( b, name, value, strlen ( value ), BSON_STRING );
+    return bson_append_string_base( b, name, value, (int)strlen ( value ), BSON_STRING );
 }
 
 MONGO_EXPORT int bson_append_symbol( bson *b, const char *name, const char *value ) {
-    return bson_append_string_base( b, name, value, strlen ( value ), BSON_SYMBOL );
+    return bson_append_string_base( b, name, value, (int)strlen ( value ), BSON_SYMBOL );
 }
 
 MONGO_EXPORT int bson_append_code( bson *b, const char *name, const char *value ) {
-    return bson_append_string_base( b, name, value, strlen ( value ), BSON_CODE );
+    return bson_append_string_base( b, name, value, (int)strlen ( value ), BSON_CODE );
 }
 
 MONGO_EXPORT int bson_append_string_n( bson *b, const char *name, const char *value, int len ) {
@@ -802,7 +802,7 @@ MONGO_EXPORT int bson_append_code_w_scope_n( bson *b, const char *name,
 }
 
 MONGO_EXPORT int bson_append_code_w_scope( bson *b, const char *name, const char *code, const bson *scope ) {
-    return bson_append_code_w_scope_n( b, name, code, strlen ( code ), scope );
+    return bson_append_code_w_scope_n( b, name, code, (int)strlen ( code ), scope );
 }
 
 MONGO_EXPORT int bson_append_binary( bson *b, const char *name, char type, const char *str, int len ) {
@@ -838,8 +838,8 @@ MONGO_EXPORT int bson_append_new_oid( bson *b, const char *name ) {
 }
 
 MONGO_EXPORT int bson_append_regex( bson *b, const char *name, const char *pattern, const char *opts ) {
-    const int plen = strlen( pattern )+1;
-    const int olen = strlen( opts )+1;
+    const int plen = (int)strlen( pattern )+1;
+    const int olen = (int)strlen( opts )+1;
     if ( bson_append_estart( b, BSON_REGEX, name, plen + olen ) == BSON_ERROR )
         return BSON_ERROR;
     if ( bson_check_string( b, pattern, plen - 1 ) == BSON_ERROR )
@@ -858,17 +858,17 @@ MONGO_EXPORT int bson_append_bson( bson *b, const char *name, const bson *bson )
 
 MONGO_EXPORT int bson_append_element( bson *b, const char *name_or_null, const bson_iterator *elem ) {
     bson_iterator next = *elem;
-    int size;
+    size_t size;
 
     bson_iterator_next( &next );
     size = next.cur - elem->cur;
 
     if ( name_or_null == NULL ) {
-        if( bson_ensure_space( b, size ) == BSON_ERROR )
+        if( bson_ensure_space( b, (int)size ) == BSON_ERROR )
             return BSON_ERROR;
-        bson_append( b, elem->cur, size );
+        bson_append( b, elem->cur, (int)size );
     } else {
-        int data_size = size - 2 - strlen( bson_iterator_key( elem ) );
+        int data_size = (int)(size - 2 - strlen( bson_iterator_key( elem ) ));
         bson_append_estart( b, elem->cur[0], name_or_null, data_size );
         bson_append( b, bson_iterator_value( elem ), data_size );
     }
@@ -919,7 +919,7 @@ MONGO_EXPORT int bson_append_start_array( bson *b, const char *name ) {
 
 MONGO_EXPORT int bson_append_finish_object( bson *b ) {
     char *start;
-    int i;
+    size_t i;
     if ( bson_ensure_space( b, 1 ) == BSON_ERROR ) return BSON_ERROR;
     bson_append_byte( b , 0 );
